feat(controllers): implement parameter validations #73

+++ src/data/usecases/ingredient/deleteIngredient-usecase.ts
@@ -1,6 +1,7 @@
 import { DeleteIngredientUseCaseInterface } from 'src/data/abstract/usecases/ingredient/deleteIngredientUseCase-interface';
 import { Ingredient } from 'src/domain/entities/ingredient';
 import { IngredientRepositoryInterface } from 'src/infra/abstract/repositories/ingredientRepository.interface';
+import { InvalidParamError } from 'src/utils/errors/invalidParam-error';
 
 export class DeleteIngredientUseCase
   implements DeleteIngredientUseCaseInterface
@@ -11,11 +12,20 @@ export class DeleteIngredientUseCase
     this.repository = repository;
   }
 
-  public execute(
+  public async execute(
     ingredientId: string,
     restaurantId: string,
   ): Promise<Ingredient> {
-    const deleted = this.repository.delete(ingredientId, restaurantId);
+    const fountEntity = await this.repository.getOne(
+      ingredientId,
+      restaurantId,
+    );
+
+    if (fountEntity === null) {
+      throw new InvalidParamError('Id');
+    }
+
+    const deleted = await this.repository.delete(ingredientId, restaurantId);
 
     return deleted;
   }
