+++ src/data/usecases/restaurant/deleteRestaurant-usecase.ts
@@ -1,6 +1,7 @@
 import { DeleteRestaurantUseCaseInterface } from 'src/data/abstract/usecases/restaurant/deleteRestaurantUseCase-interface';
 import { Restaurant } from 'src/domain/entities/restaurant';
 import { RestaurantRepositoryInterface } from 'src/infra/abstract/repositories/restaurantRepository-interface';
+import { InvalidParamError } from 'src/utils/errors/invalidParam-error';
 
 export class DeleteRestaurantUseCase
   implements DeleteRestaurantUseCaseInterface
@@ -11,8 +12,14 @@ export class DeleteRestaurantUseCase
     this.repository = repository;
   }
 
-  public execute(restaurantId: string): Promise<Restaurant> {
-    const deleted = this.repository.delete(restaurantId);
+  public async execute(restaurantId: string): Promise<Restaurant> {
+    const fountEntity = await this.repository.getOne(restaurantId);
+
+    if (fountEntity === null) {
+      throw new InvalidParamError('Id');
+    }
+
+    const deleted = await this.repository.delete(restaurantId);
 
     return deleted;
   }
